Move bond calculations into analytics module and separate out user-level calculations from other calculations.
